+++ lib/crossbeams/layout/link.rb
@@ -5,12 +5,13 @@ module Crossbeams
     # A link renderer - for rendering a link outside a form.
     class Link
       include PageNode
-      attr_reader :text, :url, :style, :behaviour, :css_class, :id, :visible
+      attr_reader :text, :url, :style, :behaviour, :css_class, :id, :visible, :text_size
 
       def initialize(options) # rubocop:disable Metrics/AbcSize
         @text      = options.fetch(:text)
         @url       = options.fetch(:url)
         @style     = options[:style] || :link
+        @colour    = options[:button_colour]
         @icon      = options[:icon]
         @behaviour = options[:behaviour] || :direct # popup window, popup dialog, modal...
         @css_class = options[:css_class] || ''
@@ -19,6 +20,7 @@ module Crossbeams
         @id        = options[:id]
         @visible   = options.fetch(:visible, true)
         @window    = options[:loading_window]
+        @text_size = options[:text_size] || '1'
         @nodes     = []
         assert_options_ok!
       end
@@ -67,24 +69,46 @@ module Crossbeams
         %(id="#{id}" )
       end
 
-      def assert_options_ok!
+      def assert_options_ok! # rubocop:disable Metrics/AbcSize, Metrics/PerceivedComplexity, Metrics/CyclomaticComplexity
+        raise ArgumentError, "Crossbeams::Layout::Link Invalid text size #{@text_size}" unless %w[1 2 3 4 5 6].include?(@text_size.to_s)
+        raise ArgumentError, 'Crossbeams::Layout::Link Only provide button colour for a `button` style' if @colour && @style != :button
+        raise ArgumentError, "Crossbeams::Layout::Link Invalid button colour option - #{@colour}" if @colour && !%i[standard red green amber blue].include?(@colour)
         return unless @window || @icon
-        raise ArgumentError, 'Crossbeams::Layout::Link you cannot have a loading window that is also a popup' if @window && %i[popup replace_dialog].include?(@behaviour)
+
+        raise ArgumentError, 'Crossbeams::Layout::Link you cannot have a loading window that is also remote or a popup' if @window && %i[popup replace_dialog remote].include?(@behaviour)
         raise ArgumentError, "Crossbeams::Layout::Link - icon #{@icon} is not a valid choice" unless @icon.nil? || Icon::ICONS.keys.include?(@icon)
         raise ArgumentError, 'Crossbeams::Layout::Link icon is not applicable for back button or loading window' if @icon && (style == :back_button || @window)
       end
 
-      def class_strings
+      def class_strings # rubocop:disable Metrics/AbcSize
         if style == :button
-          %(class="f6 link dim br2 ph3 pv2 dib white bg-silver#{user_class}")
+          col = button_colour
+          %(class="f#{button_font_size} link dim br2 ph3 pv2 dib white bg-#{col}#{user_class}")
         elsif style == :small_button
           %(class="link dim br1 ph2 dib white bg-silver#{user_class}")
         elsif style == :back_button
-          %(class="f6 link dim br2 ph3 pv2 dib white bg-dark-blue#{user_class}")
+          %(class="f#{button_font_size} link dim br2 ph3 pv2 dib white bg-dark-blue#{user_class}")
         elsif style == :action_button
-          %(class="f6 link dim br2 ph3 pv2 dib white bg-green#{user_class}")
+          %(class="f#{button_font_size} link dim br2 ph3 pv2 dib white bg-green#{user_class}")
         else
           css_class.empty? ? '' : %(class="#{css_class}")
+        end
+      end
+
+      def button_colour
+        case @colour
+        when nil, :standard
+          'silver'
+        when :green
+          'dark-green'
+        when :amber
+          'gold'
+        when :red
+          'dark-red'
+        when :blue
+          'dark-blue'
+        else
+          raise ArgumentError, "Invalid button colour - #{@colour}"
         end
       end
 
@@ -111,10 +135,13 @@ module Crossbeams
       end
 
       def behaviour_string
-        if @behaviour == :popup
+        case @behaviour
+        when :popup
           'data-popup-dialog="true"'
-        elsif @behaviour == :replace_dialog
+        when :replace_dialog
           'data-replace-dialog="true"'
+        when :remote
+          'data-remote-link="true"'
         else
           ''
         end
@@ -146,6 +173,12 @@ module Crossbeams
         return '' if @window || @prompt
 
         'data-new-page-link="true"'
+      end
+
+      def button_font_size
+        return '6' if text_size == '1'
+
+        (7 - text_size.to_i).to_s
       end
     end
   end